==> 01_Machine Learning Basics (titanic project)/02_Scripts/03_Classification Model.ipynb <==
import pandas as pd

# Load the dataset
preprcossed_csv_path = r"..\03_Outputs\Preprocessed Train.csv"
df = pd.read_csv(preprcossed_csv_path)

df.info()

#Remove Class and Features
X_train = df.drop('Survived', axis=1)
y_train = df['Survived']


from sklearn.model_selection import cross_val_score, StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
import xgboost as xgb
import numpy as np


models = {
    'Logistic Regression': LogisticRegression(max_iter=1000, random_state=42),
    'Random Forest': RandomForestClassifier(n_estimators=100, random_state=42),
    'XGBoost': xgb.XGBClassifier( eval_metric='logloss', random_state=42)
}


from sklearn.model_selection import StratifiedKFold

cv = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)


for name, model in models.items():
    print(f"Model: {name}")
    scores = cross_val_score(model, X_train, y_train, cv=cv, scoring='accuracy')
    print(scores)
    print(f"{name} Accuracy: {np.mean(scores):.4f} ± {np.std(scores):.4f}")
    print("---------------------------------------")


# Load the dataset
preprcossed_test_csv_path = r"..\03_Outputs\Preprocessed Test.csv"
df_test = pd.read_csv(preprcossed_test_csv_path)

#Remove Class and Features
X_test = df_test.drop('Survived', axis=1)
y_test = df_test['Survived']

from sklearn.metrics import accuracy_score, precision_score, recall_score, confusion_matrix

# Assuming X_train, X_test, y_train, y_test are already defined
results = {}
for name, model in models.items():
    print(f"--- {name} ---")
    model.fit(X_train, y_train)
    y_pred = model.predict(X_test)
    
    acc = accuracy_score(y_test, y_pred)
    prec = precision_score(y_test, y_pred)
    rec = recall_score(y_test, y_pred)
    cm = confusion_matrix(y_test, y_pred)
    results[name]= {"accuracy_score":acc,"precision_score":prec,"recall_score":rec,"confusion_matrix" : cm}
    print(f"Accuracy:  {acc:.4f}")
    print(f"Precision: {prec:.4f}")
    print(f"Recall:    {rec:.4f}")
    print("Confusion Matrix:")
    print(cm)
    print()


# Find the best model based on precision
best_model_precision = max(results, key=lambda x: results[x]['precision_score'])
print(f"Best model based on precision: {best_model_precision}")

# Find the best model based on recall
best_model_recall = max(results, key=lambda x: results[x]['recall_score'])
print(f"Best model based on recall: {best_model_recall}")
